refactor(aptu-coder): split metrics.rs into metrics.rs + metrics_export.rs (#1296)
